requirements.txt
streamlit
opencv-python
pyserial
numpy